Add CIDR mask to win32 netsh call for ipv6 set address
Without an explicit CIDR value, a /64 value appears to be assumed by the
OS. This causes a bogus on-link /64 route to appear when a server pushes
a smaller IPv6 address to the client.

This fix adds the mask, which works despite the CIDR value being
explicitly listed in the netsh inline help output.

Signed-off-by: Josh Cepek <josh.cepek@usa.net>